F-4384 - Cast AES key-wrap length guard operands to uint64

+++ src/wp_aes_wrap.c
@@ -349,7 +349,8 @@ static int wp_aes_wrap_update(wp_AesWrapCtx *ctx, unsigned char *out,
         ok = 0;
     }
 
-    if (ok && ((inLen > UINT32_MAX) || (outSize > UINT32_MAX))) {
+    if (ok && (((uint64_t)inLen > (uint64_t)UINT32_MAX) ||
+            ((uint64_t)outSize > (uint64_t)UINT32_MAX))) {
         ok = 0;
     }
